Moved unit test files to according namespaces.
