+++ modules/graphql_file/tests/src/Kernel/FileFieldTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\graphql_image;
+namespace Drupal\Tests\graphql_file\Kernel;
 
 use Drupal\Core\Entity\Entity\EntityViewDisplay;
 use Drupal\Core\Entity\Entity\EntityViewMode;
@@ -8,13 +8,13 @@ use Drupal\field\Entity\FieldConfig;
 use Drupal\field\Entity\FieldStorageConfig;
 use Drupal\simpletest\ContentTypeCreationTrait;
 use Drupal\simpletest\NodeCreationTrait;
-use Drupal\Tests\graphql_core\GraphQLFileTest;
+use Drupal\Tests\graphql_core\Kernel\GraphQLFileTest;
 use Drupal\user\Entity\Role;
 
 /**
  * Test file attachments.
  */
-class ImageFieldTest extends GraphQLFileTest {
+class FileFieldTest extends GraphQLFileTest {
   use NodeCreationTrait;
   use ContentTypeCreationTrait;
 
@@ -27,10 +27,8 @@ class ImageFieldTest extends GraphQLFileTest {
     'text',
     'filter',
     'file',
-    'image',
     'graphql_content',
     'graphql_file',
-    'graphql_image',
   ];
 
   /**
@@ -40,7 +38,6 @@ class ImageFieldTest extends GraphQLFileTest {
     parent::setUp();
     $this->installConfig('node');
     $this->installConfig('filter');
-    $this->installConfig('image');
     $this->installEntitySchema('node');
     $this->installSchema('node', 'node_access');
     $this->installSchema('file', 'file_usage');
@@ -58,16 +55,16 @@ class ImageFieldTest extends GraphQLFileTest {
 
 
     FieldStorageConfig::create([
-      'field_name' => 'image',
-      'type' => 'image',
+      'field_name' => 'file',
+      'type' => 'file',
       'entity_type' => 'node',
     ])->save();
 
     FieldConfig::create([
-      'field_name' => 'image',
+      'field_name' => 'file',
       'entity_type' => 'node',
       'bundle' => 'test',
-      'label' => 'Image',
+      'label' => 'File',
     ])->save();
 
     EntityViewDisplay::create([
@@ -75,30 +72,29 @@ class ImageFieldTest extends GraphQLFileTest {
       'bundle' => 'test',
       'mode' => 'graphql',
       'status' => TRUE,
-    ])->setComponent('image', ['type' => 'image'])->save();
+    ])->setComponent('file', ['type' => 'file_url_plain'])->save();
   }
 
   /**
    * Test a simple file field.
    */
-  public function testImageField() {
+  public function testFileField() {
     $a = $this->createNode([
       'title' => 'Node A',
       'type' => 'test',
     ]);
 
-    $a->image->generateSampleItems(1);
+    $a->file->generateSampleItems(1);
 
     $a->save();
 
-    $result = $this->executeQueryFile('image.gql', ['path' => '/node/' . $a->id()]);
-    $image = $result['data']['route']['node']['image'];
+    $result = $this->executeQueryFile('files.gql', ['path' => '/node/' . $a->id()]);
+    $file = $result['data']['route']['node']['file'];
 
-    $this->assertEquals($a->image->alt, $image['alt'], 'Alt text correct.');
-    $this->assertEquals($a->image->title, $image['title'], 'Title text correct.');
-    $this->assertEquals($a->image->entity->url(), $image['originalImage']['route']['internalPath'], 'Retrieve correct image url.');
-    $this->assertFalse($image['originalImage']['route']['isRouted'], 'Image urls are not routed.');
-    $this->assertFalse($image['thumbnailImage']['route']['isRouted'], 'Image urls are not routed.');
+    $this->assertEquals($a->file->entity->getSize(), $file['fileSize'], 'Retrieve correct file size.');
+    $this->assertEquals($a->file->entity->getMimeType(), $file['mimeType'], 'Retrieve correct mime type.');
+    $this->assertEquals($a->file->entity->url(), $file['entityUrl']['internalPath'], 'Retrieve correct file path.');
+    $this->assertFalse($file['entityUrl']['routed'], 'File urls are not routed.');
   }
 
 }
